+++ src/controllers/subset/index.js
@@ -1,0 +1,9 @@
+export * from './codesControl';
+export * from './descriptionControl';
+export * from './editable';
+export * from './errorsControl';
+export * from './nameControl';
+export * from './originsControl';
+export * from './restrictable';
+export * from './versionRationaleControl';
+export * from './versionsControl';
