+++ src/controllers/subset/originsControl.js
@@ -1,0 +1,15 @@
+export const originsControl = (state = {}) => ({
+
+    addOrigin(id = '') {
+        //console.debug('addOrigin', classificationId);
+
+        state._origins = [id, ...state._origins];
+    },
+
+    removeOrigin(id = '') {
+        //console.debug('removeOrigin', classificationId);
+
+        state._origins = state._origins?.filter(o => o !== id);
+        state.removeCodesWithClassificationId(id);
+    }
+});
