+++ src/controllers/subset/versionRationaleControl.js
@@ -1,0 +1,54 @@
+import { nextDefaultName, sanitize } from '../../utils';
+import { languages, subsetDraft } from '../../defaults';
+
+export const versionRationaleControl = (state = {}) => ({
+
+    addVersionRationale(versionRationale = nextDefaultName(state.versionRationale)) {
+        if (state.isEditableVersionRationale()
+            && state.versionRationale?.length < languages.filter(l => l.draft).length)
+        {
+            //console.debug('addVersionRationale', versionRationale);
+
+            state.versionRationale = [...state.versionRationale, versionRationale];
+        }
+    },
+
+    removeVersionRationaleByIndex(index) {
+        if (state.isEditableVersionRationale()
+            && index >= 0 && index < state.versionRationale?.length)
+        {
+            //console.debug('removeVersionRationaleByIndex', index);
+
+            state.versionRationale = state.currentVersion.versionRationale?.filter((item, i) => i !== index)
+        }
+    },
+
+    removeEmptyVersionRationales() {
+        if (state.isEditableVersionRationale()) {
+            //console.debug('removeEmptyVersionRationales');
+
+            state.versionRationale = state.versionRationale.filter(item => item.languageText?.length > 0);
+        }
+    },
+
+    updateVersionRationaleTextByIndex(index = -1, text = '') {
+        if (state.isEditableVersionRationale()
+            && index >= 0 && index < state.versionRationale?.length)
+        {
+            //console.debug('updateVersionRationaleTextByIndex', index, text);
+
+            state.currentVersion.versionRationale[index].languageText = sanitize(text, subsetDraft?.maxLengthDescription);
+        }
+    },
+
+    updateVersionRationaleLanguageByIndex(index = -1, lang = '') {
+        if (state.isEditableVersionRationale()
+            && index >= 0 && index < state.versionRationale?.length
+            && state.isAcceptableLanguageCode(lang))
+        {
+            //console.debug('updateVersionRationaleLanguageByIndex', index, lang);
+
+            state.currentVersion.versionRationale[index].languageCode = lang;
+        }
+    }
+});
